refactor: move schemas to shared dir
