+++ src/shared/api/commercetools/index.ts
@@ -7,8 +7,12 @@ export {
   createStoreScope,
   currencyCodeSchema,
   customerTokenScopes,
+  dateOfBirthSchema,
+  emailSchema,
   guestTokenScopes,
   mutateGetTokenInfo,
+  nameSchema,
+  passwordSchema,
   type CountryCode,
   type CurrencyCode,
   type GetTokenInfoResult,
