+++ src/shared/api/commercetools/model/email.schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const emailSchema = z.string().email();
+
+export type Email = z.infer<typeof emailSchema>;
