+++ src/shared/api/commercetools/model/index.ts
@@ -1,8 +1,12 @@
 export { ALLOWED_COUNTRIES, ALLOWED_COUNTRY_NAMES, type Country } from './allowed-countries';
 export { CTP_PROJECT_KEY } from './api-config';
 export { createScope, createStoreScope } from './create-scope';
+export { dateOfBirthSchema } from './date-of-birth.schema';
+export { emailSchema } from './email.schema';
 export { $http, httpClient } from './http-client';
 export { mutateGetTokenInfo, tokenInfoResultSchema, type GetTokenInfoResult } from './mutate-get-token-info';
+export { nameSchema } from './name.schema';
+export { passwordSchema } from './password.schema';
 export {
   countryCodeSchema,
   currencyCodeSchema,
